Split filter control into separate class
