+++ osu.Game/Screens/Select/FilterCriteria.cs
@@ -51,7 +51,7 @@ namespace osu.Game.Screens.Select
             }
         }
 
-        public FilterControl.CollectionFilter Collection;
+        public CollectionFilter Collection;
 
         public struct OptionalRange<T> : IEquatable<OptionalRange<T>>
             where T : struct
